requirements.txt
Django==4.1.13
mysqlclient==2.2.4
djangorestframework==3.14.0